Move commands into separate submodules
